Pull out Aidan's evaluation functions into a utils file, rename folder to not be specific to preweighting evaluation
