+++ asf_heat_pump_suitability/pipeline/error_analysis/error_analysis_utils.py
@@ -1,86 +1,31 @@
 """
-This script performs an error analysis on the tenure data by LSOA (Lower Layer Super Output Area).
-
-First, it loads the sample and target data. Then, it processes all LSOAs to get error metrics and differences in counts and proportions. These metrics are saved to a CSV file.
-
-Next, it creates and melts dataframes for counts and proportions differences, extracts values from these melted dataframes, and converts them to pandas dataframes.
-
-Finally, it creates boxplots and difference plots for counts and proportions differences. These plots provide a visual representation of the differences in counts and proportions by tenure for each LSOA.
-
-This script is intended to be run as a standalone script.
+Functions needed to calculate errors in the proportions of features per LSOA
 """
 
 import polars as pl
 import s3fs
-from asf_heat_pump_suitability import config
 import matplotlib.pyplot as plt
 import seaborn as sns
 import pandas as pd
 import math
 import numpy as np
-from typing import Tuple, Dict, List
-from collections import defaultdict
-from asf_heat_pump_suitability import PROJECT_DIR
-import os
 from sklearn.metrics import root_mean_squared_error, mean_absolute_error
 
+import os
+from typing import Tuple, Dict, List
+from collections import defaultdict
 
-def load_sample(sample_path):
-    """
-    Load the EPC sample from a parquet file and preprocess it.
-
-    Args:
-        sample_path (str): The path to the parquet file containing the EPC sample.
-
-    Returns:
-        sample (pl.DataFrame): The preprocessed EPC sample, with irrelevant rows filtered out and columns renamed.
-    """
-    sample = pl.read_parquet(sample_path)
-    sample = sample.filter(~pl.col("TENURE").is_in(["", "unknown"]))
-    sample = sample.select(pl.col(["lsoa", "UPRN", "TENURE"])).rename(
-        {"TENURE": "tenure", "UPRN": "id"}
-    )
-    return sample
-
-
-def load_target(target_path):
-    """
-    Load the target dataset from the census data.
-
-    Args:
-        target_path (str): The path to the Excel file containing the census data.
-
-    Returns:
-        target (pl.DataFrame): The preprocessed census data, with irrelevant rows and columns removed, and columns renamed.
-    """
-    fs = s3fs.S3FileSystem()
-    with fs.open(target_path, mode="rb") as f:
-        content = f.read()
-    target = pl.read_excel(content, sheet_name="3c", engine="calamine")
-    target = target.rename(target[2].to_dicts().pop()).slice(3)
-    target = target.rename(
-        {
-            "Area Code": "lsoa",
-            "Owned or shared ownership": "owner-occupied",
-            "Social Rented": "rental (social)",
-            "Private Rented or lives rent free": "rental (private)",
-        }
-    )
-    censored_vals = 0
-    int_cols = ["owner-occupied", "rental (private)", "rental (social)"]
-    target = target.with_columns(
-        [pl.col(int_cols).str.replace("c", f"{censored_vals}").cast(pl.Int64)]
-    ).select(pl.col(["lsoa", "owner-occupied", "rental (private)", "rental (social)"]))
-    return target
+from asf_heat_pump_suitability import PROJECT_DIR
+from asf_heat_pump_suitability import config
 
 
 def calculate_proportions(counts: dict) -> dict:
     """
     Calculate proportions for each key in a dictionary.
     Args:
-        counts (dict): dictionary with counts for each key
+            counts (dict): dictionary with counts for each key
     Returns:
-        proportions (dict): dictionary with proportions for each key
+            proportions (dict): dictionary with proportions for each key
     """
     total = sum(counts.values())
     proportions = {k: v / total for k, v in counts.items()}
@@ -91,10 +36,10 @@ def calculate_diffs(dict1: dict, dict2: dict) -> dict:
     """
     Calculate numerical differences between two dictionaries.
     Args:
-        dict1 (dict): first dictionary
-        dict2 (dict): second dictionary
+            dict1 (dict): first dictionary
+            dict2 (dict): second dictionary
     Returns:
-        differences (dict): dictionary with differences for each key
+            differences (dict): dictionary with differences for each key
     """
     keys = set(dict1) | set(dict2)
     differences = {k: dict1.get(k, 0) - dict2.get(k, 0) for k in keys}
@@ -112,14 +57,14 @@ def process_single_lsoa(
     Process a single LSOA: calculate counts, differences, RMSE, and MAE.
 
     Args:
-        sample (pl.DataFrame): The sample dataframe.
-        target (pl.DataFrame): The target dataframe.
-        test_lsoa (str): The LSOA to process.
-        lsoa (str): literally the string "lsoa"
-        target_feature_name (str): The name of the target feature.
+            sample (pl.DataFrame): The sample dataframe.
+            target (pl.DataFrame): The target dataframe.
+            test_lsoa (str): The LSOA to process.
+            lsoa (str): literally the string "lsoa"
+            target_feature_name (str): The name of the target feature.
 
     Returns:
-        Tuple[Dict, Dict, Dict, float, float]: A tuple containing the counts, differences, RMSE, and MAE.
+            Tuple[Dict, Dict, Dict, float, float]: A tuple containing the counts, differences, RMSE, and MAE.
     """
     subset = sample.filter(pl.col(lsoa) == test_lsoa)
     target_subset = target.filter(pl.col(lsoa) == test_lsoa)
@@ -212,19 +157,19 @@ def process_all_lsoas(
     Process all LSOAs and return dictionaries of RMSE, MAE, counts difference, proportions difference, and total counts.
 
     Args:
-        sample (DataFrame): The EPC sample DataFrame.
-        tenure (DataFrame): The tenure DataFrame.
-        test_lsoas (list): A list of LSOAs to test.
-        target_feature_name (str): The name of the target feature.
+            sample (DataFrame): The EPC sample DataFrame.
+            tenure (DataFrame): The tenure DataFrame.
+            test_lsoas (list): A list of LSOAs to test.
+            target_feature_name (str): The name of the target feature.
 
     Returns:
-        rmse_all_without_missing_categories (dict): A dictionary with RMSE values for each LSOA without considering missing categories.
-        mae_all_without_missing_categories (dict): A dictionary with MAE values for each LSOA without considering missing categories.
-        rmse_all_with_missing_categories (dict): A dictionary with RMSE values for each LSOA with missing categories considered.
-        mae_all_with_missing_categories (dict): A dictionary with MAE values for each LSOA with missing categories considered.
-        counts_diff_all (dict): A dictionary with counts difference for each LSOA.
-        proportions_diff_all (dict): A dictionary with proportions difference for each LSOA.
-        total_counts (dict): A dictionary with total counts for each tenure type.
+            rmse_all_without_missing_categories (dict): A dictionary with RMSE values for each LSOA without considering missing categories.
+            mae_all_without_missing_categories (dict): A dictionary with MAE values for each LSOA without considering missing categories.
+            rmse_all_with_missing_categories (dict): A dictionary with RMSE values for each LSOA with missing categories considered.
+            mae_all_with_missing_categories (dict): A dictionary with MAE values for each LSOA with missing categories considered.
+            counts_diff_all (dict): A dictionary with counts difference for each LSOA.
+            proportions_diff_all (dict): A dictionary with proportions difference for each LSOA.
+            total_counts (dict): A dictionary with total counts for each tenure type.
     """
     rmse_all_without_missing_categories = {}
     mae_all_without_missing_categories = {}
@@ -275,11 +220,11 @@ def create_error_dataframe_and_save(
     Create a DataFrame from RMSE and MAE dictionaries and save it to a CSV file.
 
     Args:
-        rmse_all_without_missing_categories (dict): A dictionary with RMSE values for each LSOA without considering missing categories.
-        mae_all_without_missing_categories (dict): A dictionary with MAE values for each LSOA without considering missing categories.
-        rmse_all_with_missing_categories (dict): A dictionary with RMSE values for each LSOA with missing categories considered.
-        mae_all_with_missing_categories (dict): A dictionary with MAE values for each LSOA with missing categories considered.
-        filename (str): The name of the CSV file to save.
+            rmse_all_without_missing_categories (dict): A dictionary with RMSE values for each LSOA without considering missing categories.
+            mae_all_without_missing_categories (dict): A dictionary with MAE values for each LSOA without considering missing categories.
+            rmse_all_with_missing_categories (dict): A dictionary with RMSE values for each LSOA with missing categories considered.
+            mae_all_with_missing_categories (dict): A dictionary with MAE values for each LSOA with missing categories considered.
+            filename (str): The name of the CSV file to save.
     """
     df_LSOA_errors = pl.DataFrame(
         {
@@ -312,13 +257,13 @@ def create_dataframe_and_melt(dictionary, column_name, target_feature_name, lsoa
     Create a DataFrame from a dictionary and melt it.
 
     Args:
-        dictionary (dict): The dictionary to convert to a DataFrame.
-        column_name (str): The name of the value column after melting.
-        target_feature_name (str): The name of the target feature.
-        lsoa (str): literally just lsoa
+            dictionary (dict): The dictionary to convert to a DataFrame.
+            column_name (str): The name of the value column after melting.
+            target_feature_name (str): The name of the target feature.
+            lsoa (str): literally just lsoa
 
     Returns:
-        pl.DataFrame: The melted DataFrame.
+            pl.DataFrame: The melted DataFrame.
     """
     df = pl.DataFrame(dictionary).transpose(include_header=True)
     df = df.rename({"column": lsoa})
@@ -333,13 +278,13 @@ def extract_values(df, column_name, target_feature_name, lsoa):
     Extract values from a DataFrame into a list of dictionaries.
 
     Args:
-        df (pl.DataFrame): The DataFrame to extract values from.
-        column_name (str): The column to extract values from.
-        target_feature_name (str): The name of the target feature.
-        lsoa (str): literally just lsoa
+            df (pl.DataFrame): The DataFrame to extract values from.
+            column_name (str): The column to extract values from.
+            target_feature_name (str): The name of the target feature.
+            lsoa (str): literally just lsoa
 
     Returns:
-        list: A list of dictionaries with the extracted values.
+            list: A list of dictionaries with the extracted values.
     """
     melted_rows = []
     for row in df.rows(named=True):
@@ -355,12 +300,12 @@ def create_boxplot(df, y, title, filename, target_feature_name, order_preference
     Create a boxplot from a DataFrame and save it to a file.
 
     Args:
-        df (pd.DataFrame): The DataFrame to create a boxplot from.
-        y (str): The column to use as the y-axis.
-        title (str): The title of the plot.
-        filename (str): The name of the file to save the plot to.
-        target_feature_name (str): The name of the target feature to be displayed on the x-axis.
-        order_preference (List[str]): A list of categories in the order of preference for plotting.
+            df (pd.DataFrame): The DataFrame to create a boxplot from.
+            y (str): The column to use as the y-axis.
+            title (str): The title of the plot.
+            filename (str): The name of the file to save the plot to.
+            target_feature_name (str): The name of the target feature to be displayed on the x-axis.
+            order_preference (List[str]): A list of categories in the order of preference for plotting.
     """
     fig, ax = plt.subplots(figsize=(10, 6))
     sns.boxplot(x=target_feature_name, y=y, data=df, ax=ax, order=order_preference)
@@ -394,15 +339,15 @@ def create_difference_plots(
     Create a grid of bar plots for each LSOA showing the differences.
 
     Args:
-        differences_all (dict): A dictionary of dictionaries where the outer keys are LSOAs and the inner keys are tenures.
-        title (str): The title for the plot.
-        filename (str): The name of the file to save the plot to.
-        label (str): The label for the y-axis (e.g., "$\Delta_{count}$" or "$\Delta_{proportion}$"
-        target_feature_name (str): The name of the target feature to be displayed on the x-axis.
-        order_preference (List[str]): A list of categories in the order of preference for plotting.
+            differences_all (dict): A dictionary of dictionaries where the outer keys are LSOAs and the inner keys are tenures.
+            title (str): The title for the plot.
+            filename (str): The name of the file to save the plot to.
+            label (str): The label for the y-axis (e.g., "$\Delta_{count}$" or "$\Delta_{proportion}$"
+            target_feature_name (str): The name of the target feature to be displayed on the x-axis.
+            order_preference (List[str]): A list of categories in the order of preference for plotting.
 
     Returns:
-        None
+            None
     """
     # List of LSOAs
     lsoas = list(differences_all.keys())
@@ -477,94 +422,3 @@ def create_difference_plots(
     if not os.path.exists(directory):
         os.makedirs(directory)
     fig.savefig(path_filename)
-
-
-if __name__ == "__main__":
-    sample_path = "s3://asf-heat-pump-suitability/outputs/epc_sample_lsoa.parquet"
-    target_path = config["data_source"]["EW_housing_characteristics_census"]
-    DELTA_COUNT = "Delta count"
-    DELTA_PROPORTION = "Delta proportion"
-    ORDER = ["owner-occupied", "rental (private)", "rental (social)"]
-    LSOA = "lsoa"
-    TARGET_FEATURE = "tenure"
-    test_lsoas = ["W01000328", "E01033942", "E01012376", "W01000527", "E01002058"]
-    number_of_lsoas = len(test_lsoas)
-    # Load the sample and target data
-    sample = load_sample(sample_path)
-    tenure = load_target(target_path)
-
-    # Process all LSOAs and get error metrics and differences
-    (
-        rmse_all_without_missing_categories,
-        mae_all_without_missing_categories,
-        rmse_all_with_missing_categories,
-        mae_all_with_missing_categories,
-        counts_diff_all,
-        proportions_diff_all,
-        total_counts,
-    ) = process_all_lsoas(sample, tenure, test_lsoas, TARGET_FEATURE)
-
-    # Save the error metrics to a CSV file
-    create_error_dataframe_and_save(
-        rmse_all_without_missing_categories,
-        mae_all_without_missing_categories,
-        rmse_all_with_missing_categories,
-        mae_all_with_missing_categories,
-        "tenure_lsoa_errors_sklearn.csv",
-    )
-
-    # Create and melt dataframes for counts and proportions differences
-    counts_df_melted = create_dataframe_and_melt(
-        counts_diff_all, DELTA_COUNT, TARGET_FEATURE, LSOA
-    )
-    proportions_df_melted = create_dataframe_and_melt(
-        proportions_diff_all, DELTA_PROPORTION, TARGET_FEATURE, LSOA
-    )
-
-    # Extract values from the melted dataframes
-    melted_rows = extract_values(counts_df_melted, DELTA_COUNT, TARGET_FEATURE, LSOA)
-    counts_df_extracted = pl.DataFrame(melted_rows)
-    proportion_melted_rows = extract_values(
-        proportions_df_melted, DELTA_PROPORTION, TARGET_FEATURE, LSOA
-    )
-    proportions_df_extracted = pl.DataFrame(proportion_melted_rows)
-
-    # Convert to pandas dataframes
-    counts_df_extracted_pd = counts_df_extracted.to_pandas()
-    proportions_df_extracted_pd = proportions_df_extracted.to_pandas()
-
-    # Create boxplots for counts and proportions differences
-    create_boxplot(
-        counts_df_extracted_pd,
-        DELTA_COUNT,
-        f"Boxplot of pre-weighted $\\Delta_{{count}}$ by Tenure for $n_{{LSOA}}={number_of_lsoas}$",
-        "delta_count_boxplot.png",
-        target_feature_name=TARGET_FEATURE,
-        order_preference=ORDER,
-    )
-    create_boxplot(
-        proportions_df_extracted_pd,
-        DELTA_PROPORTION,
-        f"Boxplot of pre-weighted $\\Delta_{{proportion}}$ by Tenure for $n_{{LSOA}}={number_of_lsoas}$",
-        "delta_proportion_boxplot.png",
-        target_feature_name=TARGET_FEATURE,
-        order_preference=ORDER,
-    )
-
-    # Create difference plots for counts and proportions differences
-    create_difference_plots(
-        counts_diff_all,
-        r"Pre-weighted $\Delta_{count}(n_{\mathrm{EPC\ properties}} - n_{\mathrm{census\ properties}})$",
-        "preweighted_delta_count_LSOAs.png",
-        "count",
-        target_feature_name=TARGET_FEATURE,
-        order_preference=ORDER,
-    )
-    create_difference_plots(
-        proportions_diff_all,
-        r"Pre-weighted $\Delta_{proportion}$",
-        "preweighted_delta_proportion_LSOAs.png",
-        "proportion",
-        target_feature_name=TARGET_FEATURE,
-        order_preference=ORDER,
-    )
